# specs/001-identidad-acceso/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Saber Uli API — Identidad y acceso
  version: 1.0.0
  summary: Identidad, acceso institucional, invitados, roles, perfil y autorización de datos.
  description: |
    Contrato de la funcionalidad 001-identidad-acceso. Es la fuente de verdad para la
    implementación (principio III) y para generar el cliente del frontend (orval).

    Convenciones:
    - Los flujos de autenticación viven en `/api/auth` (sin versión, porque la URL de retorno
      está registrada en Microsoft Entra ID). Los recursos viven en `/api/v1`.
    - Autenticación con `Authorization: Bearer <token de acceso>`. El token de renovación viaja
      solo en la cookie `su_refresh` (HttpOnly, Secure, SameSite=Strict, Path=/api/auth).
    - Errores con RFC 9457 (`application/problem+json`). El campo `type` identifica la causa con
      un URN `urn:saber-uli:problem:<slug>`; ver `components.schemas.Problem`.
    - Listas paginadas con `page` (desde 1) y `page_size` (máximo 100).
    - Todas las rutas de `/api/v1` exigen autorización de datos vigente, salvo las marcadas con
      `x-consent-exempt: true` (FR-014).
    - Las rutas marcadas con `x-requires-privileged-session: true` exigen el claim `priv`
      (ver research R-15); sin él responden 401 `reauthentication-required`.
servers:
  - url: http://localhost
    description: Desarrollo local (detrás del proxy Nginx)
...
paths:
...
  /api/v1/programs:
    get:
      tags: [programs]
      operationId: listActivePrograms
      summary: Programas activos para completar el perfil
      responses:
        '200':
          description: Programas activos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Program' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ConsentRequired' }
  /api/v1/admin/programs:
    get:
      tags: [admin, programs]
      operationId: adminListPrograms
      x-requires-privileged-session: true
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - { name: q, in: query, required: false, schema: { type: string, maxLength: 100 } }
      responses:
        '200':
          description: Programas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProgramPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin, programs]
      operationId: adminCreateProgram
      summary: Crea un programa (FR-028)
      x-requires-privileged-session: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProgramInput' }
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Program' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/admin/programs/{programId}:
    patch:
      tags: [admin, programs]
      operationId: adminUpdateProgram
      x-requires-privileged-session: true
      parameters:
        - $ref: '#/components/parameters/ProgramId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProgramPatch' }
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Program' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------------------------------------------------------- invitations
...
  /api/v1/admin/settings:
    get:
      tags: [admin]
      operationId: adminGetSettings
      x-requires-privileged-session: true
      responses:
        '200':
          description: Parámetros
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [admin]
      operationId: adminUpdateSettings
      summary: Ajusta parámetros como el plazo máximo de docentes (FR-006a)
      x-requires-privileged-session: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettingsPatch' }
      responses:
        '200':
          description: Parámetros actualizados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Settings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/admin/audit-events:
    get:
      tags: [admin]
      operationId: adminListAuditEvents
      summary: Consulta de auditoría, solo lectura (FR-035)
      x-requires-privileged-session: true
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - { name: action, in: query, required: false, schema: { type: string, maxLength: 64 } }
        - { name: actor_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: subject_user_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Eventos, del más reciente al más antiguo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEventPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: su_refresh
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    ProgramId:
      { name: programId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Unauthorized:
      description: |
        Sin sesión válida. `type`: `unauthenticated`, `session-revoked`, `account-disabled`,
        `guest-access-expired`, `guest-access-revoked` o `reauthentication-required`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConsentRequired:
      description: Falta autorización de datos vigente (`consent-required`, FR-014)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: El rol no permite esta función (`forbidden`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: No existe o está fuera del alcance del usuario (`not-found`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Conflicto con el estado actual (`conflict` u otro tipo específico)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Datos inválidos (`validation-error`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details. `detail` está en español y nunca contiene datos personales.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['urn:saber-uli:problem:guest-access-expired']
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Errores de validación por campo
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
...
    Program:
      type: object
      required: [id, code, name, campus, active]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        campus: { type: string }
        active: { type: boolean }
    ProgramInput:
      type: object
      required: [code, name, campus]
      additionalProperties: false
      properties:
        code: { type: string, pattern: '^[A-Z0-9-]{2,20}$' }
        name: { type: string, minLength: 3, maxLength: 200 }
        campus: { type: string, minLength: 2, maxLength: 100 }
        active: { type: boolean, default: true }
    ProgramPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 3, maxLength: 200 }
        campus: { type: string, minLength: 2, maxLength: 100 }
        active: { type: boolean }
    ProgramPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Program' }

    # ---------------------------------------------------------- invitaciones
...
    Settings:
      type: object
      required:
        - teacher_max_access_days
        - default_guest_access_days
        - invitation_link_ttl_days
        - sign_in_link_ttl_minutes
      properties:
        teacher_max_access_days: { type: integer, minimum: 1, maximum: 730 }
        default_guest_access_days: { type: integer, minimum: 1, maximum: 730 }
        invitation_link_ttl_days: { type: integer, minimum: 1, maximum: 30 }
        sign_in_link_ttl_minutes: { type: integer, minimum: 5, maximum: 10 }
    SettingsPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        teacher_max_access_days: { type: integer, minimum: 1, maximum: 730 }
        default_guest_access_days: { type: integer, minimum: 1, maximum: 730 }
        invitation_link_ttl_days: { type: integer, minimum: 1, maximum: 30 }
        sign_in_link_ttl_minutes: { type: integer, minimum: 5, maximum: 10 }
    AuditEvent:
      type: object
      required: [id, occurred_at, action, target_type]
      properties:
        id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
        actor_id:
          oneOf:
            - { type: string, format: uuid }
            - { type: 'null' }
        action: { type: string }
        target_type: { type: string }
        target_id: { type: string, format: uuid }
        subject_user_id: { type: string, format: uuid }
        details:
          type: object
          additionalProperties: true
    AuditEventPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AuditEvent' }